qmatmul & rename
