--- src/mase_triton/optical_compute/core.py
@@ -1,611 +1,0 @@
-"""
-Optical Transformers: https://arxiv.org/abs/2302.10360
-"""
-
-from typing import Literal
-
-import torch
-from torch import Tensor
-import triton
-import triton.language as tl
-
-from ..dtype import TORCH_DTYPE_TO_TRITON
-from ..about import PACKAGE_NAME
-
-
-def _get_autotune_configs_optical_quantize_forward_kernel():
-    block_sizes = [128, 256, 512, 1024]
-    stages = [1, 2, 3, 4]
-    configs = []
-    for bs in block_sizes:
-        for s in stages:
-            configs.append(triton.Config({"BLOCK_SIZE": bs}, num_stages=s))
-    return configs
-
-
-@triton.autotune(
-    configs=_get_autotune_configs_optical_quantize_forward_kernel(),
-    key=["n_elements"],
-    use_cuda_graph=False,
-)
-@triton.jit
-def _optical_compute_quantize_forward_kernel(
-    x_ptr,
-    output_ptr,
-    n_elements,
-    quant_levels,
-    min_val,
-    max_val,
-    lut_min,
-    seed,
-    QUANT_MODE: tl.constexpr,
-    ENABLE_LUT_MIN: tl.constexpr,
-    BLOCK_SIZE: tl.constexpr,
-    INPUT_DTYPE: tl.constexpr,
-):
-    pid = tl.program_id(axis=0)
-    offsets = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
-    load_mask = offsets < n_elements
-    x: tl.tensor = tl.load(x_ptr + offsets, mask=load_mask)
-
-    min_val = tl.full((1,), min_val, dtype=INPUT_DTYPE)
-    max_val = tl.full((1,), max_val, dtype=INPUT_DTYPE)
-    range_val = max_val - min_val
-    eps = tl.full((1,), 1e-8, dtype=INPUT_DTYPE)
-    quant_levels = tl.full((1,), quant_levels - 1, dtype=INPUT_DTYPE)
-
-    x = tl.clamp(x, min_val, max_val)
-    x = x - min_val
-    x = x / (range_val + eps)
-    x = x * quant_levels
-
-    if QUANT_MODE == "det":
-        x = tl.extra.cuda.libdevice.rint(x)
-        x = x / quant_levels
-        x = x * range_val + min_val
-    else:
-        # rand
-        bias = tl.full((1,), -0.5, dtype=INPUT_DTYPE)
-        noise = tl.rand(seed, offsets) + bias
-        x = x + noise
-        x = tl.extra.cuda.libdevice.rint(x)
-        x = x / quant_levels
-        x = x * range_val + min_val
-    if ENABLE_LUT_MIN:
-        lut_min = tl.full((1,), lut_min, dtype=INPUT_DTYPE)
-        threshold = lut_min * max_val
-        x_mask = x < threshold
-        x_mask = x_mask & (x > 0.0)
-        x = tl.where(x_mask, threshold, x)
-        threshold = lut_min * min_val.abs()
-        x_mask = x > -threshold
-        x_mask = x_mask & (x < 0.0)
-        x = tl.where(x_mask, -threshold, x)
-
-    tl.store(output_ptr + offsets, x, mask=load_mask)
-
-
-@torch.library.custom_op(
-    f"{PACKAGE_NAME}::optical_compute_quantize_fn",
-    mutates_args={},
-)
-def optical_compute_quantize_fn(
-    x: Tensor,
-    seed: int,
-    quant_levels: int,
-    min_val: float,
-    max_val: float,
-    lut_min: float | None,
-    quant_mode: str,
-) -> tuple[Tensor, int]:
-    assert x.dtype in (torch.bfloat16, torch.float16, torch.float32), f"Unsupported dtype {x.dtype}"
-    assert x.is_cuda, "Input tensor must be on CUDA device"
-    assert quant_mode in ["det", "rand"], f"Unsupported quant_mode {quant_mode}"
-    assert lut_min is None or lut_min >= 0.0, "lut_min must be non-negative"
-
-    enable_lut_min = lut_min is not None
-    lut_min = lut_min or 0.0
-
-    output = torch.empty_like(x)
-    n_elements = x.numel()
-
-    grid = lambda meta: (triton.cdiv(n_elements, meta["BLOCK_SIZE"]),)
-    _optical_compute_quantize_forward_kernel[grid](
-        x,
-        output,
-        n_elements=n_elements,
-        quant_levels=quant_levels,
-        min_val=min_val,
-        max_val=max_val,
-        lut_min=lut_min,
-        seed=seed,
-        QUANT_MODE=quant_mode,
-        ENABLE_LUT_MIN=enable_lut_min,
-        INPUT_DTYPE=TORCH_DTYPE_TO_TRITON[x.dtype],
-    )
-
-    if quant_mode == "rand":
-        seed += 1
-    return output, seed
-
-
-@optical_compute_quantize_fn.register_fake
-def _optical_quantize_forward_fn_fake(
-    x: Tensor,
-    seed: int,
-    quant_levels: int,
-    min_val: float,
-    max_val: float,
-    lut_min: float | None,
-    quant_mode: Literal["det", "rand"],
-) -> tuple[Tensor, int]:
-    output = torch.empty_like(x, dtype=x.dtype)
-    return output, seed
-
-
-def _optical_quantize_backward(ctx, *grad_outputs):
-    return grad_outputs[0], None, None, None, None, None, None
-
-
-def _optical_quantize_setup_context(ctx, inputs, output):
-    return None
-
-
-@optical_compute_quantize_fn.register_kernel("cpu")
-def _optical_quantize_forward_fn_cpu(
-    x: Tensor,
-    seed: int,
-    quant_levels: int,
-    min_val: float,
-    max_val: float,
-    lut_min: float | None,
-    quant_mode: str,
-) -> Tensor:
-    # TODO: Implement CPU kernel use torch API
-    raise NotImplementedError("CPU kernel is not implemented")
-
-
-optical_compute_quantize_fn.register_autograd(
-    _optical_quantize_backward,
-    setup_context=_optical_quantize_setup_context,
-)
-
-
-def _get_autotune_configs_forward_quantized_linear_kernel():
-    return [
-        triton.Config(
-            {"BLOCK_SIZE_M": 128, "BLOCK_SIZE_N": 256, "BLOCK_SIZE_K": 64, "GROUP_SIZE_M": 8}, num_stages=3, num_warps=8
-        ),
-        triton.Config(
-            {"BLOCK_SIZE_M": 64, "BLOCK_SIZE_N": 256, "BLOCK_SIZE_K": 32, "GROUP_SIZE_M": 8}, num_stages=4, num_warps=4
-        ),
-        triton.Config(
-            {"BLOCK_SIZE_M": 128, "BLOCK_SIZE_N": 128, "BLOCK_SIZE_K": 32, "GROUP_SIZE_M": 8}, num_stages=4, num_warps=4
-        ),
-        triton.Config(
-            {"BLOCK_SIZE_M": 128, "BLOCK_SIZE_N": 64, "BLOCK_SIZE_K": 32, "GROUP_SIZE_M": 8}, num_stages=4, num_warps=4
-        ),
-        triton.Config(
-            {"BLOCK_SIZE_M": 64, "BLOCK_SIZE_N": 128, "BLOCK_SIZE_K": 32, "GROUP_SIZE_M": 8}, num_stages=4, num_warps=4
-        ),
-        triton.Config(
-            {"BLOCK_SIZE_M": 128, "BLOCK_SIZE_N": 32, "BLOCK_SIZE_K": 32, "GROUP_SIZE_M": 8}, num_stages=4, num_warps=4
-        ),
-        triton.Config(
-            {"BLOCK_SIZE_M": 64, "BLOCK_SIZE_N": 32, "BLOCK_SIZE_K": 32, "GROUP_SIZE_M": 8}, num_stages=5, num_warps=2
-        ),
-        triton.Config(
-            {"BLOCK_SIZE_M": 32, "BLOCK_SIZE_N": 64, "BLOCK_SIZE_K": 32, "GROUP_SIZE_M": 8}, num_stages=5, num_warps=2
-        ),
-    ]
-
-
-@triton.jit
-def _noisy_quantize(
-    x: tl.tensor,
-    quant_levels,
-    min_val,
-    max_val,
-    lut_min,
-    seed,
-    INPUT_DTYPE: tl.constexpr,
-    QUANT_MODE: tl.constexpr,
-    ENABLE_LUT_MIN: tl.constexpr,
-):
-    min_val = tl.full((1,), min_val, dtype=INPUT_DTYPE)
-    max_val = tl.full((1,), max_val, dtype=INPUT_DTYPE)
-    range_val = max_val - min_val
-    eps = tl.full((1,), 1e-8, dtype=INPUT_DTYPE)
-    quant_levels = tl.full((1,), quant_levels - 1, dtype=INPUT_DTYPE)
-
-    x = tl.clamp(x, min_val, max_val)
-    x = x - min_val
-    x = x / (range_val + eps)
-    x = x * quant_levels
-
-    if QUANT_MODE == "det":
-        x = tl.extra.cuda.libdevice.rint(x)
-        x = x / quant_levels
-        x = x * range_val + min_val
-    else:
-        # rand
-        bias = tl.full((1,), -0.5, dtype=INPUT_DTYPE)
-        noise = tl.rand(seed, tl.arange(0, 1)) + bias
-        x = x + noise
-        x = tl.extra.cuda.libdevice.rint(x)
-        x = x / quant_levels
-        x = x * range_val + min_val
-    if ENABLE_LUT_MIN:
-        lut_min = tl.full((1,), lut_min, dtype=INPUT_DTYPE)
-        threshold = lut_min * max_val
-        x_mask = x < threshold
-        x_mask = x_mask & (x > 0.0)
-        x = tl.where(x_mask, threshold, x)
-        threshold = lut_min * min_val.abs()
-        x_mask = x > -threshold
-        x_mask = x_mask & (x < 0.0)
-        x = tl.where(x_mask, -threshold, x)
-    return x
-
-
-@triton.autotune(
-    configs=_get_autotune_configs_forward_quantized_linear_kernel(),
-    key=["M", "N", "K"],
-)
-@triton.jit
-def _optical_quantized_linear_forward_kernel(
-    a_ptr,
-    b_ptr,
-    c_ptr,
-    d_ptr,
-    M,
-    N,
-    K,
-    a_min,
-    a_max,
-    b_min,
-    b_max,
-    b_lut_min,
-    c_min,
-    c_max,
-    quant_levels,
-    seed,
-    stride_am,
-    stride_ak,
-    stride_bk,
-    stride_bn,
-    stride_cm,
-    stride_cn,
-    stride_d,
-    BLOCK_SIZE_M: tl.constexpr,
-    BLOCK_SIZE_N: tl.constexpr,
-    BLOCK_SIZE_K: tl.constexpr,
-    GROUP_SIZE_M: tl.constexpr,
-    ENABLE_LUT_MIN: tl.constexpr,
-    INPUT_DTYPE: tl.constexpr,
-    SKIP_QUANTIZE: tl.constexpr,
-    USE_BIAS: tl.constexpr,
-):
-
-    pid = tl.program_id(axis=0)
-    num_pid_m = tl.cdiv(M, BLOCK_SIZE_M)
-    num_pid_n = tl.cdiv(N, BLOCK_SIZE_N)
-    num_pid_in_group = GROUP_SIZE_M * num_pid_n
-    group_id = pid // num_pid_in_group
-    first_pid_m = group_id * GROUP_SIZE_M
-    group_size_m = min(num_pid_m - first_pid_m, GROUP_SIZE_M)
-    pid_m = first_pid_m + ((pid % num_pid_in_group) % group_size_m)
-    pid_n = (pid % num_pid_in_group) // group_size_m
-
-    offs_am = (pid_m * BLOCK_SIZE_M + tl.arange(0, BLOCK_SIZE_M)) % M
-    offs_bn = (pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N)) % N
-    offs_k = tl.arange(0, BLOCK_SIZE_K)
-    a_ptrs = a_ptr + (offs_am[:, None] * stride_am + offs_k[None, :] * stride_ak)
-    b_ptrs = b_ptr + (offs_k[:, None] * stride_bk + offs_bn[None, :] * stride_bn)
-
-    accumulator = tl.zeros((BLOCK_SIZE_M, BLOCK_SIZE_N), dtype=tl.float32)
-
-    for k in range(0, tl.cdiv(K, BLOCK_SIZE_K)):
-        a = tl.load(a_ptrs, mask=offs_k[None, :] < K - k * BLOCK_SIZE_K, other=0.0)
-        if not SKIP_QUANTIZE:
-            a = _noisy_quantize(
-                a,
-                quant_levels=quant_levels,
-                min_val=a_min,
-                max_val=a_max,
-                lut_min=0.0,
-                seed=seed,
-                INPUT_DTYPE=INPUT_DTYPE,
-                QUANT_MODE="det",
-                ENABLE_LUT_MIN=False,
-            )
-        b = tl.load(b_ptrs, mask=offs_k[:, None] < K - k * BLOCK_SIZE_K, other=0.0)
-        if not SKIP_QUANTIZE:
-            b = _noisy_quantize(
-                b,
-                quant_levels=quant_levels,
-                min_val=b_min,
-                max_val=b_max,
-                lut_min=b_lut_min,
-                seed=seed,
-                INPUT_DTYPE=INPUT_DTYPE,
-                QUANT_MODE="det",
-                ENABLE_LUT_MIN=ENABLE_LUT_MIN,
-            )
-        accumulator = tl.dot(a, b, accumulator)
-
-        a_ptrs += BLOCK_SIZE_K * stride_ak
-        b_ptrs += BLOCK_SIZE_K * stride_bk
-
-    c = accumulator.to(INPUT_DTYPE)
-    if not SKIP_QUANTIZE:
-        c = _noisy_quantize(
-            c,
-            quant_levels=quant_levels,
-            min_val=c_min,
-            max_val=c_max,
-            lut_min=0.0,
-            seed=seed,
-            INPUT_DTYPE=INPUT_DTYPE,
-            QUANT_MODE="rand",
-            ENABLE_LUT_MIN=False,
-        )
-    if USE_BIAS:
-        d_ptrs = d_ptr + (offs_bn * stride_d)
-        bias = tl.load(d_ptrs, mask=offs_bn < N, other=0.0)
-        c = c + bias.to(INPUT_DTYPE)
-
-    offs_cm = pid_m * BLOCK_SIZE_M + tl.arange(0, BLOCK_SIZE_M)
-    offs_cn = pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N)
-    c_ptrs = c_ptr + stride_cm * offs_cm[:, None] + stride_cn * offs_cn[None, :]
-    c_mask = (offs_cm[:, None] < M) & (offs_cn[None, :] < N)
-    tl.store(c_ptrs, c, mask=c_mask)
-
-
-@torch.library.custom_op(
-    f"{PACKAGE_NAME}::optical_compute_quantized_linear_fn",
-    mutates_args={},
-)
-def optical_compute_quantized_linear_fn(
-    x: Tensor,
-    weight: Tensor,
-    bias: Tensor | None,
-    x_min: float,
-    x_max: float,
-    w_min: float,
-    w_max: float,
-    w_lut_min: float | None,
-    o_min: float,
-    o_max: float,
-    q_levels: int,
-    q_seed: int,
-    skip_quantize: bool = False,
-) -> tuple[Tensor, int]:
-
-    assert x.dtype in (torch.bfloat16, torch.float16, torch.float32), f"Unsupported dtype {x.dtype}"
-    assert x.is_contiguous(), "Input tensor must be contiguous"
-    assert weight.ndim == 2, f"Weight tensor must be 2D, got {weight.ndim}D"
-    assert weight.dtype in (torch.bfloat16, torch.float16, torch.float32), f"Unsupported dtype {weight.dtype}"
-
-    ori_x_shape = x.size()
-    x = x.reshape(-1, ori_x_shape[-1])
-
-    if bias is not None:
-        assert bias.shape[0] == weight.shape[0], f"Bias shape {bias.shape} does not match weight shape {weight.shape}"
-
-    M, K = x.shape
-    N, K2 = weight.shape
-    assert K == K2, f"Input shape {x.shape} does not match weight shape {weight.shape}"
-
-    if bias is None:
-        output = torch.empty((M, N), device=x.device, dtype=x.dtype)
-    else:
-        output = torch.tile(bias, (M, 1)).to(x.dtype)
-
-    grid = lambda meta: (triton.cdiv(M, meta["BLOCK_SIZE_M"]) * triton.cdiv(N, meta["BLOCK_SIZE_N"]),)
-
-    _optical_quantized_linear_forward_kernel[grid](
-        x,
-        weight.T,
-        output,
-        bias if bias is not None else x,
-        M=M,
-        N=N,
-        K=K,
-        a_min=x_min,
-        a_max=x_max,
-        b_min=w_min,
-        b_max=w_max,
-        b_lut_min=w_lut_min or 0.0,
-        c_min=o_min,
-        c_max=o_max,
-        quant_levels=q_levels,
-        seed=q_seed,
-        stride_am=x.stride(0),
-        stride_ak=x.stride(1),
-        stride_bk=weight.T.stride(0),
-        stride_bn=weight.T.stride(1),
-        stride_cm=output.stride(0),
-        stride_cn=output.stride(1),
-        stride_d=bias.stride(0) if bias is not None else 0,
-        ENABLE_LUT_MIN=w_lut_min is not None,
-        INPUT_DTYPE=TORCH_DTYPE_TO_TRITON[x.dtype],
-        SKIP_QUANTIZE=skip_quantize,
-        USE_BIAS=bias is not None,
-    )
-
-    output = output.reshape(ori_x_shape[:-1] + (N,))
-    q_seed = q_seed + 1 if skip_quantize else q_seed
-    return output, q_seed
-
-
-@optical_compute_quantized_linear_fn.register_fake
-def _optical_compute_quantized_linear_fn_fake(
-    x: Tensor,
-    weight: Tensor,
-    bias: Tensor | None,
-    x_min: float,
-    x_max: float,
-    w_min: float,
-    w_max: float,
-    w_lut_min: float | None,
-    o_min: float,
-    o_max: float,
-    q_levels: int,
-    q_seed: int,
-    skip_quantize: bool = False,
-) -> tuple[Tensor, int]:
-    output = torch.empty((x.shape[0], weight.shape[0]), device=x.device, dtype=x.dtype)
-    return output, q_seed
-
-
-@optical_compute_quantized_linear_fn.register_kernel("cpu")
-def _optical_compute_quantized_linear_fn_cpu(
-    x: Tensor,
-    weight: Tensor,
-    bias: Tensor | None,
-    x_min: float,
-    x_max: float,
-    w_min: float,
-    w_max: float,
-    w_lut_min: float | None,
-    o_min: float,
-    o_max: float,
-    q_levels: int,
-    q_seed: int,
-    skip_quantize: bool = False,
-) -> tuple[Tensor, int]:
-    raise NotImplementedError("CPU kernel is not implemented")
-
-
-def _optical_quantized_linear_backward(ctx, *grad_outputs):
-    x, weight, bias = ctx.saved_tensors
-    grad_x = grad_weight = grad_bias = None
-
-    if ctx.needs_input_grad[0]:
-        grad_x = grad_outputs[0].mm(weight)
-    if ctx.needs_input_grad[1]:
-        grad_weight = grad_outputs[0].T.mm(x)
-    if bias is not None and ctx.needs_input_grad[2]:
-        grad_bias = grad_outputs[0].sum(0)
-
-    return grad_x, grad_weight, grad_bias, None, None, None, None, None, None, None, None, None, None
-
-
-def _optical_quantized_matmul_setup_context(ctx, inputs, output):
-    ctx.save_for_backward(inputs[0], inputs[1], inputs[2])
-
-
-optical_compute_quantized_linear_fn.register_autograd(
-    _optical_quantized_linear_backward,
-    setup_context=_optical_quantized_matmul_setup_context,
-)
-
-
-class OpticalTransformerFunctions:
-    """
-    OpticalTransformerFunctions provides a collection of static methods for simulating
-    optical transformer operations accelerated by Triton kernels.
-
-    - Optical Transformers: https://arxiv.org/abs/2302.10360
-    """
-
-    kernels = {
-        "optical_compute_quantize_forward_kernel": _optical_compute_quantize_forward_kernel,
-        "optical_compute_quantized_linear_forward_kernel": _optical_quantized_linear_forward_kernel,
-    }
-
-    @staticmethod
-    def optical_compute_quantize_fn(
-        x: Tensor,
-        seed: int,
-        quant_levels: int,
-        min_val: float,
-        max_val: float,
-        lut_min: float | None,
-        quant_mode: Literal["det", "rand"],
-    ) -> tuple[Tensor, int]:
-        """
-        Applies quantization to the input tensor using the specified parameters.
-
-        Args:
-            x (Tensor): The input tensor to be quantized.
-            seed (int): The random seed for reproducibility in quantization.
-            quant_levels (int): The number of quantization levels to use.
-            min_val (float): The minimum value for the quantization range.
-            max_val (float): The maximum value for the quantization range.
-            lut_min (float | None): The minimum value for the weight lookup table (LUT),
-                if applicable. If None, no LUT is used.
-            quant_mode (str): The quantization mode to use. Can be either "det" for
-                deterministic quantization or "rand" for quantization + random noise.
-
-        Returns:
-            tuple[Tensor, int]: A tuple containing the quantized tensor and an
-            integer representing the updated seed.
-        """
-        return optical_compute_quantize_fn(
-            x=x,
-            seed=seed,
-            quant_levels=quant_levels,
-            min_val=min_val,
-            max_val=max_val,
-            lut_min=lut_min,
-            quant_mode=quant_mode,
-        )
-
-    @staticmethod
-    def optical_compute_quantized_linear_fn(
-        x: Tensor,
-        weight: Tensor,
-        bias: Tensor | None,
-        x_min: float,
-        x_max: float,
-        w_min: float,
-        w_max: float,
-        w_lut_min: float | None,
-        o_min: float,
-        o_max: float,
-        q_levels: int,
-        q_seed: int,
-        skip_quantize: bool = False,
-    ) -> tuple[Tensor, int]:
-        """
-        Fused quantized linear operation: Quantize input, quantize weight with w_lut_min (if provided),
-        linear matmul, and quantize output + add noise.
-
-        Args:
-            x (Tensor): The input tensor to be transformed.
-            weight (Tensor): The weight tensor for the linear transformation.
-            bias (Tensor | None): The bias tensor for the linear transformation.
-            x_min (float): The minimum value for the input quantization range.
-            x_max (float): The maximum value for the input quantization range.
-            w_min (float): The minimum value for the weight quantization range.
-            w_max (float): The maximum value for the weight quantization range.
-            w_lut_min (float | None): The minimum value for the weight lookup table (LUT),
-                if applicable. If None, no LUT is used.
-            o_min (float): The minimum value for the output quantization range.
-            o_max (float): The maximum value for the output quantization range.
-            q_levels (int): The number of quantization levels to use.
-            q_seed (int): The random seed for reproducibility in quantization.
-            skip_quantize (bool, optional): Whether to skip quantization. Defaults to False.
-                Useful for debugging.
-
-        Returns:
-            tuple[Tensor, int]: A tuple containing the transformed tensor and an
-            integer representing the updated seed.
-        """
-        return optical_compute_quantized_linear_fn(
-            x=x,
-            weight=weight,
-            bias=bias,
-            x_min=x_min,
-            x_max=x_max,
-            w_min=w_min,
-            w_max=w_max,
-            w_lut_min=w_lut_min,
-            o_min=o_min,
-            o_max=o_max,
-            q_levels=q_levels,
-            q_seed=q_seed,
-            skip_quantize=skip_quantize,
-        )
