+++ src/mase_triton/optical_compute/core/__init__.py
@@ -1,0 +1,1 @@
+from .optical_transformer import OpticalTransformerFunctions
